pyproject.toml
[project]
name = "image-captioning-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Gautham P Raj",email = "gauthampraj2003@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "requests (>=2.32.3,<3.0.0)",
    "transformers (>=4.52.3,<5.0.0)",
    "torch (>=2.7.0,<3.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
